+++ src/app/admin/list-work/list-work.component.ts
@@ -1,5 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
+import { TypeWork } from 'src/app/models/works/works.component';
 
 @Component({
   selector: 'app-list-work',
@@ -7,23 +8,28 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./list-work.component.css']
 })
 export class ListWorkComponent implements OnInit {
-  works!:any 
+  works!:any[]
   constructor(private http: HttpClient) { }
 
   ngOnInit(): void {
     this.getWorks()
   }
   getWorks() {
-    this.http.get("http://localhost:3001/works").subscribe((data) => {
-      this.works = data
+    this.http.get<any[]>("http://localhost:3001/works").subscribe((data) => {
+     console.log(data);
+     this.works = data
+     
     })
   }
   onDel(id:number){
 
-    this.http.delete("http://localhost:3001/works/"+id).subscribe((data) => {
-           
+    const confirm = window.confirm("Bạn có chắc chắn muốn xoá không?");
+    if(confirm) { 
+      this.http.delete("http://localhost:3001/works/"+id).subscribe((data) => {
+         this.works = this.works.filter(item => item.id !== id)  
                 
     })
+    }
   }
 
 }
